# lib/api-zod/src/generated/types/chessGame.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Glo's Party Games
 * OpenAPI spec version: 0.1.0
 */
import type { ChessGameCurrentTurn } from './chessGameCurrentTurn';
import type { ChessGameStatus } from './chessGameStatus';

export interface ChessGame {
  id: number;
  hostUsername: string;
  /** @nullable */
  opponentUsername?: string | null;
  /** @nullable */
  challengerUsername?: string | null;
  status: ChessGameStatus;
  /** @nullable */
  fen?: string | null;
  /** @nullable */
  currentTurn?: ChessGameCurrentTurn;
  /** @nullable */
  winner?: string | null;
  createdAt: string;
}