# lib/api-zod/src/generated/types/chessMoveInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Glo's Party Games
 * OpenAPI spec version: 0.1.0
 */

export interface ChessMoveInput {
  move: string;
}