# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Glo's Party Games
 * OpenAPI spec version: 0.1.0
 */
import type { RoomGame } from './roomGame';
import type { RoomSettings } from './roomSettings';
import type { RoomStatus } from './roomStatus';

export interface Room {
  id: number;
  code: string;
  game: RoomGame;
  name: string;
  hostUsername: string;
  playerCount: number;
  maxPlayers: number;
  status: RoomStatus;
  settings?: RoomSettings;
  createdAt: string;
}